fix: windows dll loading
