+++ go.mod
@@ -2,4 +2,7 @@ module github.com/ploMP4/chafa-go
 
 go 1.26
 
-require github.com/ebitengine/purego v0.9.1
+require (
+	github.com/ebitengine/purego v0.9.1
+	golang.org/x/sys v0.41.0
+)
