+++ libs/BUILD_INFO.txt
@@ -1,3 +1,3 @@
 Chafa Version: 1.18
-Build Date: Mon Nov 10 20:26:58 UTC 2025
-Workflow: 19245026133
+Build Date: Tue Feb 24 08:21:14 UTC 2026
+Workflow: 22342327535
