+++ libs/CHECKSUMS.txt
@@ -1,4 +1,4 @@
-46dcb64d4de69c555785f60d907f30dbffeb20b550f4777231e581bb42573109  libs/darwin-arm64/libchafa.dylib
-7504b67767077ba04b98291084dc2583c8a5c07b740c370a7ce3b5d8ce8cfeee  libs/linux-386/libchafa.so
-d45517cd37945f85f315c98e815206948435edb3b19bb261cb24b8509c356121  libs/windows-x64/libchafa.dll
+984e6e3eb3520dc32b8d0290758fe8f4970dd438529237e1403a35908282f7a3  libs/darwin-arm64/libchafa.dylib
+b845aa26b92769c6ef7617b10948398d8d860e700b266e50188299e4456aff80  libs/linux-386/libchafa.so
+bac4168d2c07c1af63e6995cc4666b54def5f289759b4c482dcdb324df06cf47  libs/windows-x64/libchafa.dll
 5f7df37965dd03c0269dcefec685863945123f9dee5bd3c6c4014709e3bf054a  libs/linux-amd64/libchafa.so
